move interfaces of Queue and Source/Sink closer together, in preparation for using them interchangeably as template arguments

+++ runtime/deviceCode/ChannelBase.cuh
@@ -104,7 +104,7 @@ namespace Mercator  {
     // @return starting index of reserved segment.
     //
     __device__
-    unsigned int dsReserve(unsigned int nToWrite)
+    size_t dsReserve(unsigned int nToWrite)
     {
       assert(IS_BOSS());
